% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeScript.R
\name{theme_Sergio}
\alias{theme_Sergio}
\title{theme_Sergio: Theme for publications}
\usage{
theme_Sergio(base_size = 14, base_family = "sans")
}
\description{
ggplot2 theme that looks fab.
}
\examples{
plot <- ggplot(data, aes(x, y)) + geom_boxplot
plot + theme_Sergio
}
